third_party/blink/web_tests/virtual/pna-navigations-warning/external/wpt/fetch/private-network-access/iframe.tentative.https.window_include=grandparent-expected.txt

This is a testharness.js-based test.
[FAIL] public to local, grandparent navigates: failure.
  assert_equals: expected "timeout" but got "loaded"
Harness: the test ran to completion.
